Added Scaled MSE printout
I started implementing some of the choices of parameters, but then I found that the norm of the RL output is far lower than the norm of the original object and the SGP output. After scaling the RL and SGP outputs to have the same norm as the original object, the MSE and KL-divergence of the RL output improves dramatically.